fix: include response atoms in diagnostics plan
